requirements.txt
minimalmodbus
paho-mqtt
python-dotenv
<<<<<<< Updated upstream
=======
modbus-tk
pyserial
requests
>>>>>>> Stashed changes
